AI-Chatbot: refactor code
